Documentation and Edge Case Protection
Updated Documentation
Implemented edge case protection for EMG data acquisition 
- If queue for passing data between cores is full
-- Clear queue and upload most up to date data
--- Data is read from the queue twice as fast as it is added. This will only ever occur during calibration when no data in being read
Implemented edge case protection for Current Sense
- If calculated current is a negative number
-- Over write current value to zero
--- likely caused by calibration error

+++ src/SPICurrentSense/SPICurrentSense.h
@@ -39,7 +39,7 @@ class SPICurrentSense {
   // ADC response
   uint8_t result1;
   uint8_t result2;
-  uint16_t rawResult;
+  int rawResult;
 
   // Current Sense Calibration offset values
   uint16_t thumbOffset;
